processor_v2.py
import logging
import re
import urllib.parse
from bs4 import BeautifulSoup, XMLParsedAsHTMLWarning
import requests
import spacy
import os
from requests.exceptions import RequestException, Timeout, TooManyRedirects
from urllib3.exceptions import MaxRetryError
import time
from functools import wraps
from tenacity import retry, stop_after_attempt, wait_exponential
from requests.adapters import HTTPAdapter
from urllib3.util.retry import Retry
import warnings
from celery import Celery
from cachetools import TTLCache
from translator import translate_text  # Import the translation function

# Update Celery configuration
celery = Celery('tasks', broker=os.environ.get('CELERY_BROKER_URL', 'redis://localhost:6379'))
celery.conf.task_time_limit = 600  # 10 minutes
celery.conf.task_soft_time_limit = 300  # 5 minutes

cache = TTLCache(maxsize=100, ttl=3600)

logging.basicConfig(level=logging.DEBUG)

# Rest of the imports and setup code remains the same...

def remove_duplicates(blocks):
    """
    Remove duplicate blocks based on some criteria.
    """
    # Add your implementation for removing duplicate blocks here
    pass

def process_block(block, score):
    # Add your implementation for the process_block function here
    block_data = {}  # Assume this is populated with the extracted data
    
    # Add translation for text and description
    if 'text' in block_data:
        block_data['translated_text'] = translate_text(block_data['text'])
    if 'description' in block_data:
        block_data['translated_description'] = translate_text(block_data['description'])
    
    return block_data

def scrape_and_process(link):
    # Add your implementation for the scrape_and_process function here
    pass

@celery.task(bind=True)
def process_email_content(self, data):
    logging.debug(f"Starting to process email content: {data}")
    try:
        if not data or 'metadata' not in data or 'content' not in data['metadata'] or 'html' not in data['metadata']['content']:
            logging.error(f"Invalid JSON structure: {data}")
            return {"error": "Invalid JSON structure"}

        content_html = data['metadata']['content']['html']
        metadata = data['metadata']
        
        logging.debug(f"Content HTML (first 500 chars): {content_html[:500]}")
        logging.debug(f"Metadata: {metadata}")
        
        soup = BeautifulSoup(content_html, 'html.parser')

        content_blocks = []
        score = 1

        for block in soup.find_all(['div', 'table'], class_=lambda x: x and ('content-block' in x or 'hse-column' in x)):
            block_data = process_block(block, score)
            if block_data:
                if 'link' in block_data:
                    enriched_data = scrape_and_process(block_data['link'])
                    if 'error' in enriched_data:
                        logging.warning(f"Error enriching block: {enriched_data['error']}")
                    else:
                        block_data.update(enriched_data)
                
                content_blocks.append(block_data)
                score += 1

            # Update task state
            self.update_state(state='PROGRESS',
                              meta={'status': f'Processed {len(content_blocks)} blocks'})

        # Remove duplicate blocks
        unique_blocks = remove_duplicates(content_blocks)

        # Exclude social and address blocks
        final_blocks = [block for block in unique_blocks if not (
            "Follow Us" in block.get('text', '') or 
            "The Clios, 104 West 27th Street" in block.get('text', '')
        )]

        # Reassign scoring to be consecutive
        for i, block in enumerate(final_blocks, start=1):
            block['scoring'] = i

        output_json = {
            "metadata": metadata,
            "content_blocks": final_blocks
        }

        logging.debug(f"Finished processing email content: {output_json}")
        return {"result": output_json}

    except Exception as e:
        logging.error(f"Error processing email content: {str(e)}")
        return {"error": str(e)}

# The rest of the functions (clean_image_url, clean_text, scrape_and_process, etc.) remain the same...