requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
beautifulsoup4==4.12.3
gunicorn==20.1.0
requests==2.26.0
lxml==4.6.3
blinker==1.8.2
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
soupsieve==2.5
numpy==1.24.3
scipy==1.10.1
spacy==3.5.0
requests-html==0.10.0
tenacity==8.0.1
celery==5.2.7
redis==4.3.4
cachetools==5.3.0
google-cloud-translate==3.7.1